fix: fix bug Components MenuFilter and MenuNavigation follow by review of Ly Tran
