+++ src/components/common/MenuNavigation/MenuNavigation.module.scss
@@ -13,8 +13,6 @@
             li{
                 margin: 0.8rem 0;
                 a{
-                 font-size: var(--font-size);
-                 line-height: var(--line-height);
                  color:var(--text-base);
                  &:hover {
                     @apply text-primary;
